# docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Untripical API",
        "description": "API aplication Untripical",
        "contact": {
            "name": "Mateusz",
            "email": "mateusz@untripical.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Untripical server"
        }
    ],
    "paths": {
...
        "/region/{regionId}": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getRegionById",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "updateRegion",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "deleteRegion",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/region": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getAllRegions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RegionResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "addRegion",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/region/type/{type}": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getRegionByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "MOUNTAIN",
                                "COASTAL",
                                "VIEW_POINT",
                                "LAKE",
                                "FOREST",
                                "DESERT",
                                "HISTORICAL",
                                "CULTURAL",
                                "ADVENTURE",
                                "RELIGIOUS",
                                "ISLAND",
                                "DISTRICT",
                                "VOIVODESHIP",
                                "PARK"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/region/city/{closestBigCity}": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getRegionByClosestBigCity",
                "parameters": [
                    {
                        "name": "closestBigCity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReviewResponseGuideDetailsDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "guideId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usernameWhoWroteReview": {
                        "type": "string"
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "RegionRequestDTO": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "MOUNTAIN",
                            "COASTAL",
                            "VIEW_POINT",
                            "LAKE",
                            "FOREST",
                            "DESERT",
                            "HISTORICAL",
                            "CULTURAL",
                            "ADVENTURE",
                            "RELIGIOUS",
                            "ISLAND",
                            "DISTRICT",
                            "VOIVODESHIP",
                            "PARK"
                        ]
                    },
                    "closestBigCity": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "type"
                ]
            },
            "AnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "nameOfJourney": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "announcementType": {
                        "type": "string",
                        "enum": [
                            "GROUP",
                            "INDIVIDUAL"
                        ]
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "locationInfo": {
                        "type": "string"
                    },
                    "maxParticipants": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "guideAnnouncementTables": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GuideAnnouncementTableResponseDTO"
                        }
                    }
                }
            },
            "GuideAnnouncementTableResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "announcementId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "guideDetailsId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GuideDetailsResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "closestBigCity": {
                        "type": "string"
                    },
                    "specialisation": {
                        "type": "string",
                        "enum": [
                            "HISTORY",
                            "CULTURE",
                            "NATURE",
                            "MOUNTAIN_GUIDE",
                            "CITY_TOUR",
                            "FOOD_TOUR",
                            "ADVENTURE",
                            "ARCHITECTURE",
                            "RELIGIOUS",
                            "NIGHT_TOUR",
                            "BIKE_TOUR",
                            "WATER_TOUR",
                            "LOCAL_LIFE"
                        ]
                    },
                    "experienceLevel": {
                        "type": "string",
                        "enum": [
                            "BEGINNER",
                            "INTERMEDIATE",
                            "ADVANCED",
                            "EXPERT"
                        ]
                    },
                    "counterOfDidJourney": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "avgRating": {
                        "type": "number",
                        "format": "double"
                    },
                    "numberOfAnnouncements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "regionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "guideAnnouncements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GuideAnnouncementTableResponseDTO"
                        }
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReviewResponseGuideDetailsDTO"
                        }
                    }
                }
            },
            "PlaceResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "addressStreet": {
                        "type": "string"
                    },
                    "addressNumber": {
                        "type": "string"
                    },
                    "placeType": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "photoUrl": {
                        "type": "string"
                    },
                    "verificationStatus": {
                        "type": "string",
                        "enum": [
                            "WAITING_FOR_APPROVAL",
                            "APPROVED",
                            "REJECTED"
                        ]
                    },
                    "regionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "announcements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnnouncementResponseDTO"
                        }
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReviewResponsePlaceDTO"
                        }
                    }
                }
            },
            "RegionResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "MOUNTAIN",
                            "COASTAL",
                            "VIEW_POINT",
                            "LAKE",
                            "FOREST",
                            "DESERT",
                            "HISTORICAL",
                            "CULTURAL",
                            "ADVENTURE",
                            "RELIGIOUS",
                            "ISLAND",
                            "DISTRICT",
                            "VOIVODESHIP",
                            "PARK"
                        ]
                    },
                    "closestBigCity": {
                        "type": "string"
                    },
                    "places": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PlaceResponseDTO"
                        }
                    },
                    "guides": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GuideDetailsResponseDTO"
                        }
                    }
                }
            },
            "ReviewResponsePlaceDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usernameWhoWroteReview": {
                        "type": "string"
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        }
    }
}